Charmaprenderer refined

+++ src/CLGEngine/TileMap.h
@@ -6,6 +6,11 @@
 #include "CORE/Vector2.h"
 #include "MapData.h"
 
+
+// !TODO: REFINE THIS CLASS!!
+//          This class is a mess. We call it TileMap, 
+//          but it's built more like a specialized 2D Mapping classs.
+
 class TileMap {
 private:
     std::vector<CHAR_INFO> _map;
@@ -19,6 +24,6 @@ public:
     CLGEngine::Vector2<int> offset = {0, 0}; // We currently need this to grab global Tile location. Will have to set from GameManager for now.
 
     TileMap(MapData data);
-    wchar_t GetTile(CLGEngine::Vector2<float> coords);
+    wchar_t GetTile(CLGEngine::Vector2<int> coords);
     void SetMap(MapData map);
 };
